refactor(workflow-tasks): move component operations data-model into library
